+++ CHANGELOG.md
@@ -16,6 +16,8 @@ O formato é baseado em [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
   - Pequenas otimizações miscelâneas.
 - *Meds Pré-Definidas*:
   - Reintroduzidas medicações para fluxo de um certo PS específico da zona sul... #iykyk
+- *Laudos Rx*:
+  - Incluído laudos para Abdome, Colunas e Seios da Face.
 
 ## [2.15.0]
 ### Adicionado